requirements.txt
Django==2.2.3
django-ckeditor==5.7.1
django-js-asset==1.2.2
django-location-field==2.1.0
django-taggit==1.1.0
django-taggit-autosuggest==0.3.7
Pillow==6.1.0
pytz==2019.1
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
